app/Auth/oauth2.py
from datetime import datetime,timedelta,timezone
from fastapi import Depends,status,HTTPException
from fastapi.security import OAuth2PasswordBearer
from jose import JWTError,jwt
from app import models
from .. import schemas,database
from sqlalchemy.orm import Session
from ..config import settings

oauth2_scheme=OAuth2PasswordBearer(tokenUrl='login')

#secret key
#algo
#expiration time

SECRET_KEY=settings.SECRET_KEY
ALGORITHM=settings.ALGORITHM
ACCESS_TOKEN_EXPIRE_MINUTES=settings.ACCESS_TOKEN_EXPIRE_MINUTES

def create_access_token(data:dict):
    to_encode=data.copy()
    
    expire=datetime.now(timezone.utc)+timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    
    to_encode.update({"exp":expire})
    
    encoded_jwt=jwt.encode(to_encode,SECRET_KEY,algorithm=ALGORITHM)
    
    return encoded_jwt

def verify_access_token(token:str,creds_exception):
    try:
        payload=jwt.decode(token,SECRET_KEY,algorithms=[ALGORITHM])
        
        id:str =payload.get("user_id")
        
        if id is None:
            raise creds_exception
        token_data=schemas.TokenData(id=str(id))
    except JWTError:
        raise creds_exception
    
    return token_data
    

def get_current_user(token:str=Depends(oauth2_scheme),db:Session=Depends(database.get_db)):
    creds_exception=HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,detail="Couldn't validate Credentials",headers={"WWW-Authenticate":"Bearer"})
    
    token= verify_access_token(token,creds_exception)
    
    user=db.query(models.User).filter(models.User.id==token.id).first()
    
    return user
    
    